carsforsale.html & carsforsale.css
for the carforsale.css file i have updated the layout to make it easy to understand and follow through. For the carforsale.html i have edited the file paths and names to match with the css. I had to fix and update the layout too to make it easy to follow through